Some changes to help building ads.
 - Add a startTime to the context given to ads.
 - Move getAdCid to a separate function, rather than a method of AmpAd.
 - Split out ad-cid tests from amp-ad tests.
 - Fix a typo.
