+++ ads/README.md
@@ -38,6 +38,7 @@ We will provide the following information to the ad:
   - Only available on pages that load `amp-analytics`. The clientId will be null if `amp-analytics` was not loaded on the given page.
 - `window.context.pageViewId` contains a relatively low entropy id that is the same for all ads shown on a page.
 - [ad viewability](#ad-viewability)
+- `window.context.startTime` contains the time at which processing of the amp-ad element started.
 
 More information can be provided in a similar fashion if needed (Please file an issue).
