+++ builtins/amp-ad.js
@@ -16,17 +16,16 @@
 
 import {BaseElement} from '../src/base-element';
 import {IntersectionObserver} from '../src/intersection-observer';
-import {cidForOrNull} from '../src/cid';
+import {getAdCid} from '../src/ad-cid';
 import {getIframe, prefetchBootstrap} from '../src/3p-frame';
 import {isLayoutSizeDefined} from '../src/layout';
 import {listen, listenOnce, postMessage} from '../src/iframe-helper';
 import {loadPromise} from '../src/event-helper';
 import {parseUrl} from '../src/url';
 import {registerElement} from '../src/custom-element';
-import {adPrefetch, adPreconnect, clientIdScope} from '../ads/_config';
+import {adPrefetch, adPreconnect} from '../ads/_config';
 import {timer} from '../src/timer';
 import {user} from '../src/log';
-import {userNotificationManagerFor} from '../src/user-notification';
 import {viewerFor} from '../src/viewer';
 
 
@@ -234,7 +233,7 @@ export function installAd(win) {
           // now.
           loadingAdsCount--;
         }, 1000);
-        return this.getAdCid_().then(cid => {
+        return getAdCid(this).then(cid => {
           if (cid) {
             this.element.setAttribute('ampcid', cid);
           }
@@ -284,36 +283,6 @@ export function installAd(win) {
         });
       }
       return loadPromise(this.iframe_);
-    }
-
-    /**
-     * @return {!Promise<string|undefined>} A promise for a CID or undefined if
-     *     - the ad network does not request one or
-     *     - `amp-analytics` which provides the CID service was not installed.
-     * @private
-     */
-    getAdCid_() {
-      const scope = clientIdScope[this.element.getAttribute('type')];
-      const consentId = this.element.getAttribute(
-          'data-consent-notification-id');
-      if (!(scope || consentId)) {
-        return Promise.resolve();
-      }
-      return cidForOrNull(this.getWin()).then(cidService => {
-        if (!cidService) {
-          return;
-        }
-        let consent = Promise.resolve();
-        if (consentId) {
-          consent = userNotificationManagerFor(this.getWin()).then(service => {
-            return service.get(consentId);
-          });
-          if (!scope && consentId) {
-            return consent;
-          }
-        }
-        return cidService.get(scope, consent);
-      });
     }
 
     /** @override  */
